Move isInheritedComponent to parser utils
