+++ src/parser/utils.ts
@@ -1,0 +1,5 @@
+import { IndividualComponent, InheritedComponent } from './types';
+
+export function isInheritedComponent(comp: IndividualComponent | InheritedComponent) : comp is InheritedComponent {
+  return (<InheritedComponent>comp).baseComponent !== undefined;
+}
